pyproject.toml
[project]
name = "redis-semantic-router"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib",
    "seaborn>=0.12.0",
    "wordcloud>=1.9.0",
    "nltk>=3.8.0",
    "ruff>=0.7.0",
    "litellm",
    "scikit-learn",
    "sentence-transformers>=5.1.0",
    "redis-retrieval-optimizer==0.4.1"
]
